chore: refactor payload logic into separate module (#230)
* chore: refactor into payload.rs and engine.rs

* fix: fix tests

* chore: clippy

* chore: restructure RollupBoost impl

* chore: remove unused import

* chore: move engine api trait into server

* fix: fix imports

* fmt: cargo fmt
